+++ src/roboto/domain/topics/topic_reader.py
@@ -4,17 +4,21 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+from __future__ import annotations
+
 import abc
 import collections.abc
 import typing
 
-from ...time import TimeUnit
 from .operations import (
     MessagePathRepresentationMapping,
 )
 
 if typing.TYPE_CHECKING:
     import pandas  # pants: no-infer-dep
+
+
+Timestamp: typing.TypeAlias = typing.Union[int, float]
 
 
 class TopicReader(abc.ABC):
@@ -37,20 +41,16 @@ class TopicReader(abc.ABC):
     def get_data(
         self,
         message_paths_to_representations: collections.abc.Iterable[MessagePathRepresentationMapping],
-        log_time_attr_name: str,
-        log_time_unit: TimeUnit = TimeUnit.Nanoseconds,
         start_time: typing.Optional[int] = None,
         end_time: typing.Optional[int] = None,
         timestamp_message_path_representation_mapping: typing.Optional[MessagePathRepresentationMapping] = None,
-    ) -> collections.abc.Generator[dict[str, typing.Any], None, None]: ...
+    ) -> collections.abc.Generator[tuple[Timestamp, dict[str, typing.Any]], None, None]: ...
 
     @abc.abstractmethod
     def get_data_as_df(
         self,
         message_paths_to_representations: collections.abc.Iterable[MessagePathRepresentationMapping],
-        log_time_attr_name: str,
-        log_time_unit: TimeUnit = TimeUnit.Nanoseconds,
         start_time: typing.Optional[int] = None,
         end_time: typing.Optional[int] = None,
         timestamp_message_path_representation_mapping: typing.Optional[MessagePathRepresentationMapping] = None,
-    ) -> "pandas.DataFrame": ...
+    ) -> tuple[pandas.Series, pandas.DataFrame]: ...